fix: account detail 500 (Decimal/float mismatch) + HM Land Registry API URL
- Change purchase_price from Float to Numeric(18,2) so template arithmetic
  with Decimal balance doesn't raise TypeError
- Update HM Land Registry to current endpoint:
  /data/ppi/transaction-record.json (old /linked-data/ppd returns 404)
- Parse RFC 2822 transactionDate for correct sort order
- Tested against SE13 5GA: returns £4.1M last-sold price

+++ app/models/account.py
@@ -74,7 +74,7 @@ class Account(Base):
 
     # Real estate / physical asset fields
     property_address: Mapped[str | None] = mapped_column(String(500))
-    purchase_price: Mapped[float | None] = mapped_column(Float)
+    purchase_price: Mapped[Decimal | None] = mapped_column(Numeric(18, 2))
     purchase_date: Mapped[datetime | None] = mapped_column(DateTime)
     linked_mortgage_account_id: Mapped[int | None] = mapped_column(Integer)
